# openapi.yaml
openapi: 3.0.2
info:
  description: Moov Image Cash Letter (ICL) implements an HTTP API for creating, parsing and validating ImageCashLetter files.
  version: v1
  title: ImageCashLetter API
  contact:
    url: https://github.com/cardcorp/imagecashletter
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
servers:
  - url: http://localhost:8083
    description: Local development
...
paths:
...
  /files:
    get:
      tags: ['Image Cash Letter Files']
      summary: Get ICL Files
      operationId: getICLFiles
      security:
        - bearerAuth: []
        - cookieAuth: []
      parameters:
        - name: X-Request-ID
          in: header
          description: Optional Request ID allows application developer to trace requests through the systems logs
          example: rs4f9915
          schema:
            type: string
      responses:
        '200':
          description: A list of File objects
          headers:
            X-Total-Count:
              description: The total number of Originators
              schema:
                type: integer
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ICLFiles'
  /files/create:
    post:
      tags: ['Image Cash Letter Files']
      summary: Create File
      operationId: createICLFile
      security:
        - bearerAuth: []
        - cookieAuth: []
      parameters:
        - name: X-Request-ID
          in: header
          description: Optional Request ID allows application developer to trace requests through the systems logs
          example: rs4f9915
          schema:
            type: string
        - name: X-IDempotency-Key
          in: header
          description: Idempotent key in the header which expires after 24 hours. These strings should contain enough entropy for to not collide with each other in your requests.
          example: a4f88150
          required: false
          schema:
            type: string
      requestBody:
        description: Content of the ImageCashLetter file (in json or raw text)
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateICLFile'
          text/plain:
            schema:
              $ref: '#/components/schemas/RawICLFile'
      responses:
        '201':
          description: A JSON object containing a new File
          headers:
            Location:
              description: The location of the new resource
              schema:
                type: string
                format: uri
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ICLFile'
        '400':
          description: "Invalid File Header Object"
          content:
            application/json:
              schema:
                $ref: 'https://raw.githubusercontent.com/moov-io/base/master/api/common.yaml#/components/schemas/Error'
  /files/{fileID}:
    get:
      tags: ['Image Cash Letter Files']
      summary: Retrieve a file
      description: Retrieves the details of an existing File. You need only supply the unique File identifier that was returned upon creation.
      operationId: getICLFileByID
      security:
        - bearerAuth: []
        - cookieAuth: []
      parameters:
        - name: X-Request-ID
          in: header
          description: Optional Request ID allows application developer to trace requests through the systems logs
          example: rs4f9915
          schema:
            type: string
        - name: fileID
          in: path
          description: File ID
          required: true
          schema:
            type: string
            example: 3f2d23ee214
      responses:
        '200':
          description: A File object for the supplied ID
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ICLFile'
        '404':
          description: A resource with the specified ID was not found
    post:
      tags: ['Image Cash Letter Files']
      summary: Updates FileHeader
      description: Updates the specified File Header by setting the values of the parameters passed. Any parameters not provided will be left unchanged.
      operationId: updateICLFile
      security:
        - bearerAuth: []
        - cookieAuth: []
      parameters:
        - name: X-Request-ID
          in: header
          description: Optional Request ID allows application developer to trace requests through the systems logs
          example: rs4f9915
          schema:
            type: string
        - name: X-IDempotency-Key
          in: header
          description: Idempotent key in the header which expires after 24 hours. These strings should contain enough entropy for to not collide with each other in your requests.
          example: a4f88150
          required: false
          schema:
            type: string
        - name: fileID
          in: path
          description: File ID
          required: true
          schema:
            type: string
            example: 3f2d23ee214
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ICLFileHeader'
      responses:
        '201':
          description: A JSON object containing a new File
          headers:
            Location:
              description: The location of the new resource
              schema:
                type: string
                format: uri
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ICLFile'
        '400':
          description: "Invalid File Header Object"
          content:
            application/json:
              schema:
                $ref: 'https://raw.githubusercontent.com/moov-io/base/master/api/common.yaml#/components/schemas/Error'
    delete:
      tags: ['Image Cash Letter Files']
      summary: Delete file
      description: Permanently deletes a File and associated CashLetters and Bundles. It cannot be undone.
      operationId: deleteICLFile
      security:
        - bearerAuth: []
        - cookieAuth: []
      parameters:
        - name: fileID
          in: path
          description: File ID
          required: true
          schema:
            type: string
            example: 3f2d23ee214
        - name: X-Request-ID
          in: header
          description: Optional Request ID allows application developer to trace requests through the systems logs
          example: rs4f9915
          schema:
            type: string
      responses:
        '200':
          description: Permanently deleted File.
        '400':
          description: A problem was encountered deleting the file, check errors.
  /files/{fileID}/contents:
    get:
      tags: ['Image Cash Letter Files']
      summary: Get file contents
      description: Assembles the existing file (Cash Letters, Bundles and Controls) records, computes sequence numbers and totals. Returns plaintext file.
      operationId: getICLFileContents
      security:
        - bearerAuth: []
        - cookieAuth: []
      parameters:
        - name: X-Request-ID
          in: header
          description: Optional Request ID allows application developer to trace requests through the systems logs
          example: rs4f9915
          schema:
            type: string
        - name: fileID
          in: path
          description: File ID
          required: true
          schema:
            type: string
            example: 3f2d23ee214
      responses:
        '200':
          description: File built successfully without errors.
          content:
            text/plain:
              schema:
                $ref: '#/components/schemas/RawICLFile'
  /files/{fileID}/validate:
    get:
      tags: ['Image Cash Letter Files']
      summary: Validate file
      description: Validates the existing file. You need only supply the unique File identifier that was returned upon creation.
      operationId: validateICLFile
      security:
        - bearerAuth: []
        - cookieAuth: []
      parameters:
        - name: X-Request-ID
          in: header
          description: Optional Request ID allows application developer to trace requests through the systems logs
          example: rs4f9915
          schema:
            type: string
        - name: fileID
          in: path
          description: File ID
          required: true
          schema:
            type: string
            example: 3f2d23ee214
      responses:
        '200':
          description: File validated successfully without errors.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ICLFile'
        '400':
          description: Validation failed. Check response for errors
  /files/{fileID}/cashLetters:
    post:
      tags: ['Image Cash Letter Files']
      summary: Add CashLetter to File
      operationId: addICLToFile
      security:
        - bearerAuth: []
        - cookieAuth: []
      parameters:
        - name: X-Request-ID
          in: header
          description: Optional Request ID allows application developer to trace requests through the systems logs
          example: rs4f9915
          schema:
            type: string
        - name: X-IDempotency-Key
          in: header
          description: Idempotent key in the header which expires after 24 hours. These strings should contain enough entropy for to not collide with each other in your requests.
          example: a4f88150
          required: false
          schema:
            type: string
        - name: fileID
          in: path
          description: File ID
          required: true
          schema:
            type: string
            example: 3f2d23ee214
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CashLetter'
      responses:
        '200':
          description: CashLetter added to File
  /files/{fileID}/cashLetters/{cashLetterID}:
    delete:
      tags: ['Image Cash Letter Files']
      summary: Delete a CashLetter from a File
      operationId: deleteICLFromFile
      security:
        - bearerAuth: []
        - cookieAuth: []
      parameters:
        - name: X-Request-ID
          in: header
          description: Optional Request ID allows application developer to trace requests through the systems logs
          example: rs4f9915
          schema:
            type: string
        - name: fileID
          in: path
          description: File ID
          required: true
          schema:
            type: string
            example: 3f2d23ee214
        - name: cashLetterID
          in: path
          description: CashLetter ID
          required: true
          schema:
            type: string
            example: 45758063
      responses:
        '200':
          description: CashLetter deleted
        '404':
          description: CashLetter or File not found
components:
  schemas:
    CreateICLFile:
      properties:
        ID:
          type: string
          description: File ID
          example: 3f2d23ee214
        fileHeader:
          $ref: '#/components/schemas/ICLFileHeader'
        cashLetters:
          type: array
          items:
            $ref: '#/components/schemas/CashLetter'
        bundles:
          type: array
          items:
            $ref: '#/components/schemas/Bundle'
        fileControl:
          $ref: '#/components/schemas/ICLFileControl'
      required:
        - fileHeader
    ICLFile:
      properties:
        ID:
          type: string
          description: File ID
          example: 3f2d23ee214
        fileHeader:
          $ref: '#/components/schemas/ICLFileHeader'
        cashLetters:
          type: array
          items:
            $ref: '#/components/schemas/CashLetter'
        bundles:
          type: array
          items:
            $ref: '#/components/schemas/Bundle'
        fileControl:
          $ref: '#/components/schemas/ICLFileControl'
    ICLFileHeader:
      properties:
        ID:
          description: FileHeader ID
          type: string
          example: d1e26288
        standardLevel:
          type: string
          maxLength: 2
          enum:
            - 03
            - 30
            - 35
          description: |
            StandardLevel identifies the standard level of the file.

            * `03` - DSTU X9.37 - 2003
            * `30` - X9.100-187-2008
            * `35` - X9.100-187-2013 and 2016
          example: '35'
        testFileIndicator:
          type: string
          maxLength: 1
          enum:
            - T
            - P
          description: |
            TestFileIndicator dentifies whether the file is a test or production file.

            * `T` - Test File
            * `P` - Production File
          example: 'P'
        immediateDestination:
          type: string
          maxLength: 9
          description: |
            ImmediateDestination is the routing and transit number of the Federal Reserve Bank (FRB)
            or receiver to which the file is being sent.
          example: '231380102'
        immediateOrigin:
          type: string
          maxLength: 9
          description: |
            ImmediateOrigin is the routing and transit number of the Federal Reserve Bank (FRB)
            or originator from which the file is being sent.
          example: '201982859'
        fileCreationDate:
          type: string
          format: 'date-time'
          description: FileCreationDate is the date that the immediate origin institution creates the file.
          example: 2018-10-03T00:00:00Z
        fileCreationTime:
          type: string
          maxLength: 4
          description: |
            FileCreationTime is the time the immediate origin institution creates the file.
            (Format - hhmm, where - hh hour, mm minute)
          example: '1310'
        resendIndicator:
          type: string
          maxLength: 1
          description: ResendIndicator Indicates whether the file has been previously transmitted. (Y - Yes, N - No)
          example: 'Y'
        immediateDestinationName:
          type: string
          maxLength: 18
          description: ImmediateDestinationName Identifies the short name of the institution that receives the file.
          example: 'ABC Bank'
        immediateOriginName:
          type: string
          maxLength: 18
          description: immediateOriginName identifies the short name of the institution that sends the file.
          example: 'Bank One'
        fileIDModifier:
          type: string
          maxLength: 1
          description: |
            FileIDModifier is a code that permits multiple files, created on the same date, same time and between the
            same institutions, to be distinguished one from another. If all of the following fields in a previous file are
            equal to the same fields in this file: FileHeader ImmediateDestination, ImmediateOrigin, FileCreationDate, and
            FileCreationTime, it must be defined.
          example: 'A'
        countryCode:
          type: string
          maxLength: 2
          description: |
            CountryCode is a 2-character code as approved by the International Organization for Standardization (ISO) used
            to identify the country in which the payer bank is located.
          example: 'US'
        userField:
          type: string
          maxLength: 4
          description: UserField identifies a field used at the discretion of users of the standard.
          example: 'ZZZZ'
        companionDocumentIndicator:
          type: string
          maxLength: 1
          description: |
            CompanionDocumentIndicator identifies a field used to indicate the Companion Document being used.
            Shall be present only under clearing arrangements. Companion Document usage and values
            defined by clearing arrangements.
            Values:
            0–9 Reserved for United States use
            A–J Reserved for Canadian use
            Other - as defined by clearing arrangements.
          example: '1'
      required:
        - standardLevel
        - testFileIndicator
        - resendIndicator
        - immediateDestination
        - immediateOrigin
        - fileCreationDate
        - fileCreationTime
    ICLFileControl:
      properties:
        ID:
          type: string
          description: FileControl ID
          example: 'd1e26288'
        cashLetterCount:
          type: integer
          description: CashLetterCount identifies the total number of cash letters within the file.
          example: 1
        totalRecordCount:
          type: integer
          description: TotalRecordCount identifies the total number of records of all types sent in the file, including the FileControl.
          example: 1
        totalItemCount:
          type: integer
          description: totalItemCount identifies the total number of Items sent within the file.
          example: 5
        fileTotalAmount:
          type: integer
          description: FileTotalAmount identifies the total Item amount of the complete file.
          example: 100
        immediateOriginContactName:
          type: string
          maxLength: 14
          description: immediateOriginContactName identifies contact at the institution that creates the file.
          example: 'John Smith'
        immediateOriginContactPhoneNumber:
          type: string
          maxLength: 10
          description: ImmediateOriginContactPhoneNumber identifies the phone number of the contact at the institution that creates the file.
          example: '555-555-1212'
        creditTotalIndicator:
          type: string
          maxLength: 1
          enum:
            - 0
            - 1
          description: |
            CreditTotalIndicator is a code that indicates whether Credits Items are included in this record’s totals. If so they will be included in TotalItemCount and FileTotalAmount.
            TotalRecordCount includes all records of all types regardless of the value of this field.
            * ` ` - No Credit Items
            * `0` - Credit Items are not included in totals
            * `1` - Credit Items are included in totals
          example: 1
      required:
        - cashLetterCount
        - totalRecordCount
        - totalItemCount
        - fileTotalAmount
    RawICLFile:
      type: string
      description: A plaintext ImageCashLetter file. See our test files or examples for a complete ICL/X9 file.
      example: |
        0135T061000146026073150202010230911NWave Money        Wave Money        1      1
        9900000100000012000000010000000000010000                        0
    ICLFiles:
      type: array
      items:
        $ref: '#/components/schemas/ICLFile'
    CashLetter:
      properties:
        cashLetterHeader:
          $ref: '#/components/schemas/CashLetterHeader'
        # Current Implementation: CreditItem(s) Precede Check(s) - CreditItem(s) outside the leading Bundle
        # and Within the First Cash Letter.  Please adjust reader and writer, and openAPI for your specific clearing arrangement
        # implementation or contact MOOV for your particular implementation.
        creditItems:
          type: array
          items:
            $ref: '#/components/schemas/CreditItem'
        bundles:
          type: array
          items:
            $ref: '#/components/schemas/Bundle'
        routingNumberSummary:
          type: array
          items:
            $ref: '#/components/schemas/RoutingNumberSummary'
        cashLetterControl:
          $ref: '#/components/schemas/CashLetterControl'
...
    CashLetterControl:
      properties:
        ID:
          type: string
          description: CashLetterControl ID
          example: 'd1e26288'
        cashLetterBundleCount:
          type: integer
          description: CashLetterBundleCount identifies the total number of bundles within the cash letter.
          example: 1
        cashLetterItemsCount:
          type: integer
          description: CashLetterItemsCount identifies the total number of items within the cash letter.
          example: 1
        cashLetterTotalAmount:
          type: integer
          description: CashLetterTotalAmount identifies the total dollar value of all item amounts within the cash letter.
          example: 100
        cashLetterImagesCount:
          type: integer
          description: CashLetterImagesCount identifies the total number of ImageViewDetail(s) within the CashLetter.
          example: 1
        eCEInstitutionName:
          type: string
          maxLength: 18
          description: ECEInstitutionName identifies the short name of the institution that creates the CashLetterControl.
          example: "Citadel"
        settlementDate:
          type: string
          format: 'date-time'
          description: SettlementDate identifies the date that the institution that creates the cash letter expects settlement.
          example: 2018-10-03T00:00:00Z
        creditTotalIndicator:
          maxLength: 1
          enum:
            - 0
            - 1
          description: |
            CreditTotalIndicator is a code that indicates whether Credits Items are included in this record’s totals.
            If so they will be included in TotalItemCount and FileTotalAmount. TotalRecordCount includes all records of
            all types regardless of the value of this field.
            * ` ` - No Credit Items
            * `0` - Credit Items are not included in totals
            * `1` - Credit Items are included in totals
          example: 1
      required:
        - cashLetterItemsCount
        - cashLetterTotalAmount
        - settlementDate
...
    Bundle:
      properties:
        bundleHeader:
          $ref: '#/components/schemas/BundleHeader'
        checks:
          type: array
          items:
            $ref: '#/components/schemas/Checks'
        returns:
          type: array
          items:
            $ref: '#/components/schemas/Returns'
        bundleControl:
          $ref: '#/components/schemas/BundleControl'
...
    RoutingNumberSummary:
      properties:
        ID:
          type: string
          description: RoutingNumberSummary ID
          example: d1e26288
        cashLetterRoutingNumber:
          type: string
          maxLength: 9
          description: CashLetterRoutingNumber is a number that identifies a given payor bank within a cash letter
            containing one or more payor banks.
          example: '231380104'
        routingNumberTotalAmount:
          type: integer
          description: RoutingNumberTotalAmount is the total amount for all Check Records associated with the
            Payor bank routing number designated in the Routing Number within Cash Letter.
          example: 10000
        routingNumberItemCount:
          type: integer
          description: RoutingNumberItemCount is the the total number of all Check Records associated with the
            payor bank routing number designated in the Routing Number within Cash Letter.
          example: 10
        userField:
          type: string
          maxLength: 24
          description: UserField identifies a field used at the discretion of users of the standard.
          example: "B1234567891012345"
      required:
        - cashLetterRoutingNumber
...